Add files via upload
This file was an introduction to binary classification and the confusion matrix. When we have binary outcomes and a binary classification, we can be right in two different ways and we can be wrong in two different ways. If the true condition is the thing that we are trying to identify, in the case of the radar image, the true bombers. And we correctly classify an event as positive, this is known as a True Positive. On the other hand, if we correctly identify a condition as nothing to be worried about, this the seagull or some other blotch on our radar that does not have any particular meaning, we are classifying something as negative and it is a True Negative. The idea that we're trying to communicate is that our binary classification system involves something that generates ratings. Ratings that can be turned into rankings, to which we can apply a threshold. Then everything above the threshold is going to be declared positive, and everything below the threshold is going to be declared negative.